Add LikePost model for tracking post likes
This commit introduces the LikePost model, which is used to track user likes on posts. The model consists of two fields: 'post_id' to store the identifier of the liked post and 'username' to store the username of the user who liked the post. The '__str__' method is overridden to provide a human-readable representation of the model instance by returning the username of the user who liked the post.

Changes introduced:
- Added the LikePost model to models.py
- Defined fields 'post_id' and 'username'
- Implemented '__str__' method for human-readable representation